+++ tigergrav/gravity.hpp
@@ -6,22 +6,22 @@
 
 
 struct multipole_info {
-	multipole<float> m;
-	vect<float> x;
-	float r;
+	multipole<double> m;
+	vect<double> x;
+	double r;
 };
 
 struct multi_src {
-	multipole<float > m;
-	vect<float> x;
+	multipole<double > m;
+	vect<double> x;
 };
 
 
 std::uint64_t gravity_direct(std::vector<force> &g, const std::vector<vect<float>> &x, std::vector<source> &y);
 std::uint64_t gravity_direct_multipole(std::vector<force> &g, const std::vector<vect<float>> &x, std::vector<multi_src> &y);
-std::uint64_t gravity_indirect_multipole(expansion<double>&, const vect<float> &x, std::vector<multi_src> &y);
+std::uint64_t gravity_indirect_multipole(expansion<double>&, const vect<double> &x, std::vector<multi_src> &y);
 std::uint64_t gravity_ewald(std::vector<force> &g, const std::vector<vect<float>> &x, std::vector<source> &y);
 std::uint64_t gravity_indirect_ewald(expansion<double> &L, const vect<float> &x, std::vector<source> &y);
-float ewald_near_separation(const vect<float> x);
-float ewald_far_separation(const vect<float> x);
+double ewald_near_separation(const vect<double> x);
+double ewald_far_separation(const vect<double> x);
 void init_ewald();
